pyproject.toml
[tool.poetry]
name = "line-api-utils"
version = "0.1.0"
description = ""
authors = ["MysMon <mysmon.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
rich = "^13.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
